+++ CHANGELOG.md
@@ -1,3 +1,23 @@
 # CHANGELOG
 
-## 1.0.0 (2020-10-22)
+## [1.1.0](https://github.com/vuejs/vue-next/compare/v1.0.0...v1.1.0) (2020-10-27)
+
+
+### Bug Fixes
+
+* notify wathcer when delete object of '$data' ([e03f374](https://github.com/ayan0312/mvvm/commit/e03f374198985dfab63f426765a1545f1e38111d))
+
+### Code Refactoring
+
+* lifecycle ([41a8225](https://github.com/ayan0312/mvvm/commit/41a82258341afa358db04b693a6ff7256160342c))
+* replace def with Proxy ([7057f0d](https://github.com/ayan0312/mvvm/commit/7057f0d956d57d4136fd30f3a16c4765dc4b150c))
+
+### Features
+
+* v-style ([6d8a9e9](https://github.com/ayan0312/mvvm/commit/6d8a9e99c6fc1cf97ccad9ffe2310076ae4e8088))
+* event ([d853270](https://github.com/ayan0312/mvvm/commit/d8532706df1eb18ccf579d9664ebf9d273e4a2f6))
+* component
+* slot
+* v-for
+* nextTick
+* refs